requirements.txt
langchain_openai
langchain
python-dotenv
langchain_community
streamlit
langchain_core
bs4
arxiv
pypdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
ipykernel
langchain_groq
huggingface_hub